added service layer

+++ src/main/java/com/sprint/project/patientAppointment/Repository/AppointmentRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sprint.project.patientAppointment.Entity.AppointmentEntity;
 
-public interface AppointmentRepository extends JpaRepository<AppointmentEntity, Long> {
+public interface AppointmentRepository extends JpaRepository<AppointmentEntity, Integer> {
 
 }
