+++ src/main/java/com/sprint/project/patientAppointment/Repository/PatientRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sprint.project.patientAppointment.Entity.PatientEntity;
 
-public interface PatientRepository extends JpaRepository<PatientEntity, Long> {
+public interface PatientRepository extends JpaRepository<PatientEntity, Integer> {
 
 }
